re2: Add support for ancient versions of libre2
Ubuntu 16.04 ships libre2 2016-02-01. Back then, libre2 lacked a few niceties
that we're using. Now, while Ubuntu 16.04 shouldn't really be used out there
in this day and age, it is. We could just upgrade all uses of Ubuntu 16.04 to
a newer version, or we could add compatibility for them. Turns out, simply
adding these changes is the lesser of the two evils.

Good thing upstream (the ruby gem, not libre2) also wants to retain ancient
compat, which is where this change was originally incubated.